Update intro.md
The Screens created from the built-in Screen Templates have sample data that can you can replace manually or semi-automatically. Having sample data in the Screen is a good way to see how the Screen is designed and to get inspired for developing your app.
"sample data that can you can replace"?